Changed database model again
Django doesn't support composite PKs so we created a new field for the many-to-many table person_campaign (and dropped the existing PK from when we designed and set up the Postgres database). It also set campaign_id_campaign as a OnetoOneField so we changed it to a ForeignKey. Django did not detect these changes when running makemigration so we had to manually edit the migration files.